Move table creation to the API.
