+++ src/article/ArticleAPI.js
@@ -8,6 +8,7 @@ import renderEntity from './shared/renderEntity'
 import TranslateableModel from './models/TranslateableModel'
 import { REQUIRED_PROPERTIES } from './ArticleConstants'
 import TableEditingAPI from './shared/TableEditingAPI'
+import { generateTable } from './shared/tableHelpers'
 
 // TODO: this should come from configuration
 const COLLECTIONS = {
@@ -397,6 +398,7 @@ export default class ArticleAPI extends EditorAPI {
   // TODO: can we improve this?
   // 1. Here we would need a transaction on archive level, creating assets, plus placing them inside the article body.
   // 2. It would be interesting to use a more generic interface maybe even JATS
+  // TODO: Use JATS snippet for figure creation
   _insertFigures (files) {
     const articleSession = this.articleSession
     let LAST = files.length - 1
@@ -433,6 +435,25 @@ export default class ArticleAPI extends EditorAPI {
         }
       })
     })
+  }
+
+  // TODO: Use JATS snippet for table creation
+  _createTableFigure (tx, params) {
+    let caption = tx.createElement('caption').append(
+      tx.createElement('p')
+    )
+    let permission = tx.create({
+      type: 'permission'
+    })
+    let table = generateTable(tx, params.rows, params.columns)
+    let tableFigure = tx.create({
+      type: 'table-figure',
+      caption: caption.id,
+      content: table.id,
+      permission: permission.id
+    })
+
+    return tableFigure
   }
 
   _insertFootnote (item, footnotes) {
